+++ CHANGELOG.md
@@ -16,6 +16,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 ### Added
 - ids parameter to Transaction.query
 - ids parameter to Transfer.query
+- hiddenFields parameter to Boleto.pdf
+- ourNumber attribute to Boleto
 
 ### Fixed
 - Fixed params alterations on queries